Extract ExtensionsModel
Plugins, skills, marketplaces, MCP servers with credential storage and
device-flow auth, and the MCP input/authorization prompts move behind a
feature model that owns the auth coordinator and the debounced refresh
task. Six MCP/extension dispatcher cases collapse to one ingest call.
The credential-binding statics move with their domain; their test call
sites now exercise ExtensionsModel directly.

New tests cover refresh debouncing, the auth-required surface, the
undeliverable-input guard, the runtime credential filter, inertness, and
the publication bridge.
